First monolithic commit
Creates the basic structure of the frontend. More to come.
Merge remote-tracking branch 'origin/master'


BW-01

Updating class diagram
[DOC][CLASS][BW-01]

These diagrams ilustrate the relationship between the class "Home", which represents the landing page of the application, and the sub pages. Every class delivers a web page with a list of items, such as services or skills, hence the "header" and "subtitle" attributes common to all classes. Idividual items will be clickable, leading to a detail page.
Merge pull request #1 from theocarranza/BW-01

[DOC][CLASS][BW-01]

These diagrams ilustrate the relationship between the class "Home", which represents the landing page of the application, and the sub pages. Every class delivers a web page with a list of items, such as services or skills, hence the "header" and "subtitle" attributes common to all classes. Idividual items will be clickable, leading to a detail page.